+++ packages/vtable/examples/list/list.ts
@@ -19,8 +19,16 @@ export function createTable() {
   const records = generatePersons(1000000);
   const columns: VTable.ColumnsDefine = [
     {
+      field: '',
+      title: '行号',
+      width: 80,
+      fieldFormat(data, col, row, table) {
+        return row - 1;
+      }
+    },
+    {
       field: 'id',
-      title: 'ID ff',
+      title: 'ID',
       width: '1%',
       minWidth: 200,
       sort: true
@@ -303,10 +311,8 @@ export function createTable() {
     columns,
     tooltip: {
       isShowOverflowTextTooltip: true
-    }
-    // maintainedDataCount: 200,
-    // frozenColCount: 2,
-    // rightFrozenColumnCount: 1
+    },
+    frozenColCount: 1
     // autoWrapText: true,
     // heightMode: 'autoHeight',
     // widthMode: 'adaptive'