+++ packages/vtable/src/ts-types/list-table/define/index.ts
@@ -1,19 +1,21 @@
 import type { Either } from '../../../tools/helper';
 import type { ICompositeColumnBodyDefine } from './Composite-define';
 import type { IChartColumnBodyDefine } from './chart-define';
+import type { ICheckboxColumnBodyDefine, ICheckboxHeaderDefine } from './checkbox-define';
 import type { IImageColumnBodyDefine, IImageHeaderDefine } from './image-define';
 import type { ILinkColumnBodyDefine, ILinkHeaderDefine } from './link-define';
 import type { ITextColumnBodyDefine, ITextHeaderDefine } from './multilinetext-define';
 import type { IProgressbarColumnBodyDefine } from './progressbar-define';
 import type { ISparklineColumnBodyDefine } from './sparkline-define';
 
-export type HeaderDefine = IImageHeaderDefine | ILinkHeaderDefine | ITextHeaderDefine;
+export type HeaderDefine = IImageHeaderDefine | ILinkHeaderDefine | ICheckboxHeaderDefine | ITextHeaderDefine;
 
 export type ColumnBodyDefine =
   | ILinkColumnBodyDefine
   | IImageColumnBodyDefine
   | ISparklineColumnBodyDefine
   | IProgressbarColumnBodyDefine
+  | ICheckboxColumnBodyDefine
   | IChartColumnBodyDefine
   | ITextColumnBodyDefine;
 export type TextColumnDefine = ITextColumnBodyDefine & HeaderDefine;
@@ -21,6 +23,7 @@ export type LinkColumnDefine = ILinkColumnBodyDefine & HeaderDefine;
 export type ImageColumnDefine = IImageColumnBodyDefine & HeaderDefine;
 export type SparklineColumnDefine = ISparklineColumnBodyDefine & HeaderDefine;
 export type ProgressbarColumnDefine = IProgressbarColumnBodyDefine & HeaderDefine;
+export type CheckboxColumnDefine = ICheckboxColumnBodyDefine & HeaderDefine;
 export type ChartColumnDefine = IChartColumnBodyDefine & HeaderDefine;
 export type CompositeColumnDefine = ICompositeColumnBodyDefine & HeaderDefine;
 // export type GroupColumnDefine = IChartColumnBodyDefine & HeaderDefine;
@@ -33,6 +36,7 @@ export type ColumnDefine = Either<
   | ImageColumnDefine
   | SparklineColumnDefine
   | ProgressbarColumnDefine
+  | CheckboxColumnDefine
   | ChartColumnDefine
   | TextColumnDefine
   | CompositeColumnDefine,